Api/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core
azure-functions
azure-cosmos
azure-storage-blob
azure-ai-formrecognizer
azure-ai-documentintelligence
reportlab
openai
pandas
python-docx
pypdf